Not pretty ShoppingCart is working

+++ WildBearAdventuresMVC/WildBear/Models/DTOs/ShoppingCartDto.cs
@@ -9,16 +9,16 @@ public class ShoppingCartDto
     public DateTime createdDate { get; set; }
     public string cultureCode { get; set; }
     public object[] customProperties { get; set; }
-    public float discount { get; set; }
-    public float discountTotal { get; set; }
+    public decimal discount { get; set; }
+    public decimal discountTotal { get; set; }
     public string id { get; set; }
     public object note { get; set; }
     public Orderline[] orderLines { get; set; }
-    public float orderTotal { get; set; }
-    public float paymentTotal { get; set; }
-    public float shippingTotal { get; set; }
-    public float subTotal { get; set; }
-    public float tax { get; set; }
+    public decimal orderTotal { get; set; }
+    public decimal paymentTotal { get; set; }
+    public decimal shippingTotal { get; set; }
+    public decimal subTotal { get; set; }
+    public decimal tax { get; set; }
 }
 
 public class Billingcurrency
@@ -29,18 +29,18 @@ public class Billingcurrency
 
 public class Orderline
 {
-    public float discount { get; set; }
+    public decimal discount { get; set; }
     public object[] discounts { get; set; }
     public string id { get; set; }
     public Orderlineproperty[] orderLineProperties { get; set; }
-    public float price { get; set; }
+    public decimal price { get; set; }
     public string productName { get; set; }
     public int quantity { get; set; }
     public string sku { get; set; }
-    public float tax { get; set; }
-    public float taxRate { get; set; }
-    public float total { get; set; }
-    public float unitDiscount { get; set; }
+    public decimal tax { get; set; }
+    public decimal taxRate { get; set; }
+    public decimal total { get; set; }
+    public decimal unitDiscount { get; set; }
     public string variantSku { get; set; }
 }
